Enhancements for addresint field:
Integration to the Google Maps Places API which provides a search bar for easily finding and populating address information.
Ability to customize the way address is displayed, either on an item level, or by usage of a display format template.
Embed interactive maps for panning and zooming, as well as better responsiveness