Issue with using an alias in MySQL WHERE clause
Now use the expression from the valgrp select expression to build where
clause. Maybe put that scan in a function or use other code if you have
in your framework. The '_valgrp_in_'
substitute is not needed anymore,
but sequence must be value,text,valgrp or value,text,image,valgrp.