Improve logic that displays separators in phone numbers field
In the previous setup the country code - area code separator was displayed only when the country code was displayed. This created an issue when we didn't want to include country code but needed a prefix before the area code, for example:

(212) 999-8800

This commit changes the logic used to display the separators. The country code - area code separator will now display when either of those codes is displayed.